Crate google_firebase1_beta1

This documentation was generated from Firebase Management crate version 1.0.9+20190703, where 20190703 is the exact revision of the firebase:v1beta1 schema built by the mako code generator v1.0.9.

Everything else about the Firebase Management v1_beta1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

available projects

list

operations

get

projects

add firebase, android apps create, android apps get, android apps get config, android apps list, android apps patch, android apps sha create, android apps sha delete, android apps sha list, available locations list, default location finalize, get, get admin sdk config, ios apps create, ios apps get, ios apps get config, ios apps list, ios apps patch, list, patch, search apps, web apps create, web apps get, web apps get config, web apps list and web apps patch

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().add_firebase(...).doit()
let r = hub.projects().default_location_finalize(...).doit()
let r = hub.operations().get(...).doit()
let r = hub.projects().ios_apps_create(...).doit()
let r = hub.projects().android_apps_create(...).doit()
let r = hub.projects().web_apps_create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-firebase1_beta1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_firebase1_beta1 as firebase1_beta1;
use firebase1_beta1::AddFirebaseRequest;
use firebase1_beta1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use firebase1_beta1::FirebaseManagement;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = FirebaseManagement::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = AddFirebaseRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().add_firebase(req, "project")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AddFirebaseRequest
All fields are required.

AdminSdkConfig
There is no detailed description.

AndroidApp
Details of a Firebase App for Android.

AndroidAppConfig
Configuration metadata of a single Firebase App for Android.

AvailableProjectListCall
Returns a list of [Google Cloud Platform (GCP) Projects] (https://cloud.google.com/resource-manager/reference/rest/v1/projects) that are available to have Firebase resources added to them.

A GCP Project will only be returned if:

AvailableProjectMethods
A builder providing access to all methods supported on availableProject resources. It is not used directly, but through the FirebaseManagement hub.

Chunk

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DefaultResources
The default resources associated with a FirebaseProject.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FinalizeDefaultLocationRequest
There is no detailed description.

FirebaseAppInfo
A high-level summary of an App.

FirebaseManagement
Central instance to access all FirebaseManagement related resource activities

FirebaseProject
A FirebaseProject is the top-level Firebase entity. It is the container for Firebase Apps, Firebase Hosting sites, storage systems (Firebase Realtime Database, Cloud Firestore, Cloud Storage buckets) and other Firebase and Google Cloud Platform (GCP) resources.

You can create a new FirebaseProject in the Firebase console, or you can add Firebase resources to an existing GCP Project by calling AddFirebase.

A FirebaseProject is actually also a GCP Project, so a FirebaseProject uses underlying GCP identifiers (most importantly, projectId) as its own for easy interop with GCP APIs.

IosApp
Details of a Firebase App for iOS.

IosAppConfig
Configuration metadata of a single Firebase App for iOS.

JsonServerError
A utility type which can decode a server response that indicates error

ListAndroidAppsResponse
There is no detailed description.

ListAvailableLocationsResponse
There is no detailed description.

ListAvailableProjectsResponse
There is no detailed description.

ListFirebaseProjectsResponse
There is no detailed description.

ListIosAppsResponse
There is no detailed description.

ListShaCertificatesResponse
There is no detailed description.

ListWebAppsResponse
There is no detailed description.

Location
A GCP resource location that can be selected for a Project.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the FirebaseManagement hub.

ProjectAddFirebaseCall
Adds Firebase resources to the specified existing [Google Cloud Platform (GCP) Project] (https://cloud.google.com/resource-manager/reference/rest/v1/projects).

Since a FirebaseProject is actually also a GCP Project, a FirebaseProject uses underlying GCP identifiers (most importantly, the projectId) as its own for easy interop with GCP APIs.

The result of this call is an Operation. Poll the Operation to track the provisioning process by calling GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation succeeded, its response is set to a FirebaseProject; if the Operation failed, its error is set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

This method does not modify any billing account information on the underlying GCP Project.

All fields listed in the request body are required.

To call AddFirebase, a member must be an Editor or Owner for the existing GCP Project. Service accounts cannot call AddFirebase.

ProjectAndroidAppCreateCall
Requests that a new AndroidApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectAndroidAppGetCall
Gets the AndroidApp identified by the specified resource name.

ProjectAndroidAppGetConfigCall
Gets the configuration artifact associated with the specified AndroidApp.

ProjectAndroidAppListCall
Lists each AndroidApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectAndroidAppPatchCall
Updates the attributes of the AndroidApp identified by the specified resource name.

ProjectAndroidAppShaCreateCall
Adds a SHA certificate to the specified AndroidApp.

ProjectAndroidAppShaDeleteCall
Removes a SHA certificate from the specified AndroidApp.

ProjectAndroidAppShaListCall
Returns the list of SHA-1 and SHA-256 certificates for the specified AndroidApp.

ProjectAvailableLocationListCall
Returns a list of valid Google Cloud Platform (GCP) resource locations for the specified Project (including a FirebaseProject).

The default GCP resource location of a project defines the geographical location where project resources, such as Cloud Firestore, will be provisioned by default.

The returned list are the available GCP resource locations.

This call checks for any location restrictions for the specified Project and, thus, might return a subset of all possible GCP resource locations. To list all GCP resource locations (regardless of any restrictions), call the endpoint without specifying a projectId (that is, /v1beta1/{parent=projects/-}/listAvailableLocations).

To call ListAvailableLocations with a specified project, a member must be at minimum a Viewer of the project. Calls without a specified project do not require any specific project permissions.

ProjectDefaultLocationFinalizeCall
Sets the default Google Cloud Platform (GCP) resource location for the specified FirebaseProject.

This method creates an App Engine application with a default Cloud Storage bucket, located in the specified location_id. This location must be one of the available GCP resource locations.

After the default GCP resource location is finalized, or if it was already set, it cannot be changed. The default GCP resource location for the specified FirebaseProject might already be set because either the GCP Project already has an App Engine application or FinalizeDefaultLocation was previously called with a specified location_id. Any new calls to FinalizeDefaultLocation with a different specified location_id will return a 409 error.

The result of this call is an Operation, which can be used to track the provisioning process. The response type of the Operation is google.protobuf.Empty.

The Operation can be polled by its name using GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation has succeeded, its response will be set to a google.protobuf.Empty; if the Operation has failed, its error will be set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

All fields listed in the request body are required.

To call FinalizeDefaultLocation, a member must be an Owner of the project.

ProjectGetAdminSdkConfigCall
Gets the configuration artifact used by servers to simplify initialization.

Typically, this configuration is used with the Firebase Admin SDK initializeApp command.

ProjectGetCall
Gets the FirebaseProject identified by the specified resource name.

ProjectInfo
A reference to a Google Cloud Platform (GCP) Project.

ProjectIosAppCreateCall
Requests that a new IosApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectIosAppGetCall
Gets the IosApp identified by the specified resource name.

ProjectIosAppGetConfigCall
Gets the configuration artifact associated with the specified IosApp.

ProjectIosAppListCall
Lists each IosApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectIosAppPatchCall
Updates the attributes of the IosApp identified by the specified resource name.

ProjectListCall
Lists each FirebaseProject accessible to the caller.

The elements are returned in no particular order, but they will be a consistent view of the Projects when additional requests are made with a pageToken.

This method is eventually consistent with Project mutations, which means newly provisioned Projects and recent modifications to existing Projects might not be reflected in the set of Projects. The list will include only ACTIVE Projects.

Use GetFirebaseProject for consistent reads as well as for additional Project details.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the FirebaseManagement hub.

ProjectPatchCall
Updates the attributes of the FirebaseProject identified by the specified resource name.

All query parameters are required.

ProjectSearchAppCall
A convenience method that lists all available Apps for the specified FirebaseProject.

Typically, interaction with an App should be done using the platform-specific service, but some tool use-cases require a summary of all known Apps (such as for App selector interfaces).

ProjectWebAppCreateCall
Requests that a new WebApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectWebAppGetCall
Gets the WebApp identified by the specified resource name.

ProjectWebAppGetConfigCall
Gets the configuration artifact associated with the specified WebApp.

ProjectWebAppListCall
Lists each WebApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectWebAppPatchCall
Updates the attributes of the WebApp identified by the specified resource name.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SearchFirebaseAppsResponse
There is no detailed description.

ServerError

ServerMessage

ShaCertificate
A SHA-1 or SHA-256 certificate associated with the AndroidApp.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

WebApp
Details of a Firebase App for the web.

WebAppConfig
Configuration metadata of a single Firebase App for the web.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.