Module google_firebase1_beta1::api

Structs

AddFirebaseRequest
All fields are required.

AddGoogleAnalyticsRequest
There is no detailed description.

AdminSdkConfig
There is no detailed description.

AnalyticsDetails
There is no detailed description.

AnalyticsProperty
Details of a Google Analytics property

AndroidApp
Details of a Firebase App for Android.

AndroidAppConfig
Configuration metadata of a single Firebase App for Android.

AvailableProjectListCall
Lists each [Google Cloud Platform (GCP) Project] (https://cloud.google.com/resource-manager/reference/rest/v1/projects) that can have Firebase resources added to it. A Project will only be listed if: - The caller has sufficient Google IAM permissions to call AddFirebase. - The Project is not already a FirebaseProject. - The Project is not in an Organization which has policies that prevent Firebase resources from being added.

AvailableProjectMethods
A builder providing access to all methods supported on availableProject resources. It is not used directly, but through the FirebaseManagement hub.

DefaultResources
The default resources associated with the Project.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

FinalizeDefaultLocationRequest
There is no detailed description.

FirebaseAppInfo
A high-level summary of an App.

FirebaseManagement
Central instance to access all FirebaseManagement related resource activities

FirebaseProject
A FirebaseProject is the top-level Firebase entity. It is the container for Firebase Apps, Firebase Hosting sites, storage systems (Firebase Realtime Database, Cloud Firestore, Cloud Storage buckets), and other Firebase and Google Cloud Platform (GCP) resources. You create a FirebaseProject by calling AddFirebase and specifying an existing GCP Project. This adds Firebase resources to the existing GCP Project. Since a FirebaseProject is actually also a GCP Project, a FirebaseProject has the same underlying GCP identifiers (projectNumber and projectId). This allows for easy interop with Google APIs.

IosApp
Details of a Firebase App for iOS.

IosAppConfig
Configuration metadata of a single Firebase App for iOS.

ListAndroidAppsResponse
There is no detailed description.

ListAvailableLocationsResponse
There is no detailed description.

ListAvailableProjectsResponse
There is no detailed description.

ListFirebaseProjectsResponse
There is no detailed description.

ListIosAppsResponse
There is no detailed description.

ListShaCertificatesResponse
There is no detailed description.

ListWebAppsResponse
There is no detailed description.

Location
A GCP resource location that can be selected for a FirebaseProject.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the FirebaseManagement hub.

ProjectAddFirebaseCall
Adds Firebase resources to the specified existing [Google Cloud Platform (GCP) Project] (https://cloud.google.com/resource-manager/reference/rest/v1/projects). Since a FirebaseProject is actually also a GCP Project, a FirebaseProject has the same underlying GCP identifiers (projectNumber and projectId). This allows for easy interop with Google APIs. The result of this call is an Operation. Poll the Operation to track the provisioning process by calling GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation succeeded, its response is set to a FirebaseProject; if the Operation failed, its error is set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation. This method does not modify any billing account information on the underlying GCP Project. To call AddFirebase, a project member or service account must have the following permissions (the IAM roles of Editor and Owner contain these permissions): firebase.projects.update, resourcemanager.projects.get, serviceusage.services.enable, and serviceusage.services.get.

ProjectAddGoogleAnalyticCall
Links the specified FirebaseProject with an existing Google Analytics account. Using this call, you can either: - Specify an analyticsAccountId to provision a new Google Analytics property within the specified account and associate the new property with the FirebaseProject. - Specify an existing analyticsPropertyId to associate the property with the FirebaseProject. Note that when you call AddGoogleAnalytics: 1. The first check determines if any existing data streams in the Google Analytics property correspond to any existing Firebase Apps in the FirebaseProject (based on the packageName or bundleId associated with the data stream). Then, as applicable, the data streams and apps are linked. Note that this auto-linking only applies to AndroidApps and IosApps. 2. If no corresponding data streams are found for the Firebase Apps, new data streams are provisioned in the Google Analytics property for each of the Firebase Apps. Note that a new data stream is always provisioned for a Web App even if it was previously associated with a data stream in the Analytics property. Learn more about the hierarchy and structure of Google Analytics accounts in the Analytics documentation. The result of this call is an Operation. Poll the Operation to track the provisioning process by calling GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation succeeded, its response is set to an AnalyticsDetails; if the Operation failed, its error is set to a google.rpc.Status. To call AddGoogleAnalytics, a project member must be an Owner for the existing FirebaseProject and have the Edit permission for the Google Analytics account. If the FirebaseProject already has Google Analytics enabled, and you call AddGoogleAnalytics using an analyticsPropertyId that's different from the currently associated property, then the call will fail. Analytics may have already been enabled in the Firebase console or by specifying timeZone and regionCode in the call to AddFirebase.

ProjectAndroidAppCreateCall
Requests the creation of a new AndroidApp in the specified FirebaseProject. The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectAndroidAppGetCall
Gets the specified AndroidApp.

ProjectAndroidAppGetConfigCall
Gets the configuration artifact associated with the specified AndroidApp.

ProjectAndroidAppListCall
Lists each AndroidApp associated with the specified FirebaseProject. The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectAndroidAppPatchCall
Updates the attributes of the specified AndroidApp.

ProjectAndroidAppShaCreateCall
Adds a ShaCertificate to the specified AndroidApp.

ProjectAndroidAppShaDeleteCall
Removes a ShaCertificate from the specified AndroidApp.

ProjectAndroidAppShaListCall
Lists the SHA-1 and SHA-256 certificates for the specified AndroidApp.

ProjectAvailableLocationListCall
Lists the valid Google Cloud Platform (GCP) resource locations for the specified Project (including a FirebaseProject). One of these locations can be selected as the Project's default GCP resource location, which is the geographical location where the Project's resources, such as Cloud Firestore, will be provisioned by default. However, if the default GCP resource location has already been set for the Project, then this setting cannot be changed. This call checks for any possible location restrictions for the specified Project and, thus, might return a subset of all possible GCP resource locations. To list all GCP resource locations (regardless of any restrictions), call the endpoint without specifying a unique project identifier (that is, /v1beta1/{parent=projects/-}/listAvailableLocations). To call ListAvailableLocations with a specified project, a member must be at minimum a Viewer of the Project. Calls without a specified project do not require any specific project permissions.

ProjectDefaultLocationFinalizeCall
Sets the default Google Cloud Platform (GCP) resource location for the specified FirebaseProject. This method creates an App Engine application with a default Cloud Storage bucket, located in the specified locationId. This location must be one of the available GCP resource locations. After the default GCP resource location is finalized, or if it was already set, it cannot be changed. The default GCP resource location for the specified FirebaseProject might already be set because either the underlying GCP Project already has an App Engine application or FinalizeDefaultLocation was previously called with a specified locationId. Any new calls to FinalizeDefaultLocation with a different specified locationId will return a 409 error. The result of this call is an Operation, which can be used to track the provisioning process. The response type of the Operation is google.protobuf.Empty. The Operation can be polled by its name using GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation has succeeded, its response will be set to a google.protobuf.Empty; if the Operation has failed, its error will be set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation. All fields listed in the request body are required. To call FinalizeDefaultLocation, a member must be an Owner of the Project.

ProjectGetAdminSdkConfigCall
Gets the configuration artifact associated with the specified FirebaseProject, which can be used by servers to simplify initialization. Typically, this configuration is used with the Firebase Admin SDK initializeApp command.

ProjectGetAnalyticsDetailCall
Gets the Google Analytics details currently associated with the specified FirebaseProject. If the FirebaseProject is not yet linked to Google Analytics, then the response to GetAnalyticsDetails is NOT_FOUND.

ProjectGetCall
Gets the specified FirebaseProject.

ProjectInfo
A reference to a Google Cloud Platform (GCP) Project.

ProjectIosAppCreateCall
Requests the creation of a new IosApp in the specified FirebaseProject. The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectIosAppGetCall
Gets the specified IosApp.

ProjectIosAppGetConfigCall
Gets the configuration artifact associated with the specified IosApp.

ProjectIosAppListCall
Lists each IosApp associated with the specified FirebaseProject. The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectIosAppPatchCall
Updates the attributes of the specified IosApp.

ProjectListCall
Lists each FirebaseProject accessible to the caller. The elements are returned in no particular order, but they will be a consistent view of the Projects when additional requests are made with a pageToken. This method is eventually consistent with Project mutations, which means newly provisioned Projects and recent modifications to existing Projects might not be reflected in the set of Projects. The list will include only ACTIVE Projects. Use GetFirebaseProject for consistent reads as well as for additional Project details.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the FirebaseManagement hub.

ProjectPatchCall
Updates the attributes of the specified FirebaseProject. All query parameters are required.

ProjectRemoveAnalyticCall
Unlinks the specified FirebaseProject from its Google Analytics account. This call removes the association of the specified FirebaseProject with its current Google Analytics property. However, this call does not delete the Google Analytics resources, such as the Google Analytics property or any data streams. These resources may be re-associated later to the FirebaseProject by calling AddGoogleAnalytics and specifying the same analyticsPropertyId. For Android Apps and iOS Apps, this call re-links data streams with their corresponding apps. However, for Web Apps, this call provisions a new data stream for each Web App. To call RemoveAnalytics, a project member must be an Owner for the FirebaseProject.

ProjectSearchAppCall
Lists all available Apps for the specified FirebaseProject. This is a convenience method. Typically, interaction with an App should be done using the platform-specific service, but some tool use-cases require a summary of all known Apps (such as for App selector interfaces).

ProjectWebAppCreateCall
Requests the creation of a new WebApp in the specified FirebaseProject. The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

ProjectWebAppGetCall
Gets the specified WebApp.

ProjectWebAppGetConfigCall
Gets the configuration artifact associated with the specified WebApp.

ProjectWebAppListCall
Lists each WebApp associated with the specified FirebaseProject. The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

ProjectWebAppPatchCall
Updates the attributes of the specified WebApp.

RemoveAnalyticsRequest
There is no detailed description.

SearchFirebaseAppsResponse
There is no detailed description.

ShaCertificate
A SHA-1 or SHA-256 certificate associated with the AndroidApp.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

StreamMapping
A mapping of a Firebase App to a Google Analytics data stream

WebApp
Details of a Firebase App for the web.

WebAppConfig
Configuration metadata of a single Firebase App for the web.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.