Struct google_firebase1_beta1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the FirebaseManagement hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_firebase1_beta1 as firebase1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use firebase1_beta1::FirebaseManagement;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = FirebaseManagement::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_firebase(...)`, `android_apps_create(...)`, `android_apps_get(...)`, `android_apps_get_config(...)`, `android_apps_list(...)`, `android_apps_patch(...)`, `android_apps_sha_create(...)`, `android_apps_sha_delete(...)`, `android_apps_sha_list(...)`, `available_locations_list(...)`, `default_location_finalize(...)`, `get(...)`, `get_admin_sdk_config(...)`, `ios_apps_create(...)`, `ios_apps_get(...)`, `ios_apps_get_config(...)`, `ios_apps_list(...)`, `ios_apps_patch(...)`, `list(...)`, `patch(...)`, `search_apps(...)`, `web_apps_create(...)`, `web_apps_get(...)`, `web_apps_get_config(...)`, `web_apps_list(...)` and `web_apps_patch(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn ios_apps_patch(
&self,
request: IosApp,
name: &str
) -> ProjectIosAppPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the attributes of the IosApp identified by the specified resource name.

Arguments

request - No description provided.

name - The fully qualified resource name of the App, in the format:
projects/projectId/iosApps/appId

pub fn default_location_finalize(
&self,
request: FinalizeDefaultLocationRequest,
parent: &str
) -> ProjectDefaultLocationFinalizeCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the default Google Cloud Platform (GCP) resource location for the specified FirebaseProject.

This method creates an App Engine application with a default Cloud Storage bucket, located in the specified location_id. This location must be one of the available GCP resource locations.

After the default GCP resource location is finalized, or if it was already set, it cannot be changed. The default GCP resource location for the specified FirebaseProject might already be set because either the GCP Project already has an App Engine application or FinalizeDefaultLocation was previously called with a specified location_id. Any new calls to FinalizeDefaultLocation with a different specified location_id will return a 409 error.

The result of this call is an Operation, which can be used to track the provisioning process. The response type of the Operation is google.protobuf.Empty.

The Operation can be polled by its name using GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation has succeeded, its response will be set to a google.protobuf.Empty; if the Operation has failed, its error will be set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

All fields listed in the request body are required.

To call FinalizeDefaultLocation, a member must be an Owner of the project.

Arguments

request - No description provided.

parent - The resource name of the Project for which the default GCP resource location will be set, in the format:
projects/projectId

pub fn android_apps_get(&self, name: &str) -> ProjectAndroidAppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the AndroidApp identified by the specified resource name.

Arguments

name - The fully qualified resource name of the App, in the format:
projects/projectId/androidApps/appId
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/androidApps/appId

pub fn android_apps_sha_list(
&self,
parent: &str
) -> ProjectAndroidAppShaListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the list of SHA-1 and SHA-256 certificates for the specified AndroidApp.

Arguments

parent - The parent App for which to list SHA certificates, in the format:
projects/projectId/androidApps/appId
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/androidApps/appId

pub fn android_apps_patch(
&self,
request: AndroidApp,
name: &str
) -> ProjectAndroidAppPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the attributes of the AndroidApp identified by the specified resource name.

Arguments

request - No description provided.

name - The fully qualified resource name of the App, in the format:
projects/projectId/androidApps/appId

pub fn get(&self, name: &str) -> ProjectGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the FirebaseProject identified by the specified resource name.

Arguments

name - The fully qualified resource name of the Project, in the format:
projects/projectId

pub fn list(&self) -> ProjectListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists each FirebaseProject accessible to the caller.

The elements are returned in no particular order, but they will be a consistent view of the Projects when additional requests are made with a pageToken.

This method is eventually consistent with Project mutations, which means newly provisioned Projects and recent modifications to existing Projects might not be reflected in the set of Projects. The list will include only ACTIVE Projects.

Use GetFirebaseProject for consistent reads as well as for additional Project details.

pub fn add_firebase(
&self,
request: AddFirebaseRequest,
project: &str
) -> ProjectAddFirebaseCall<'a, C, A>

Create a builder to help you perform the following task:

Adds Firebase resources to the specified existing [Google Cloud Platform (GCP) Project] (https://cloud.google.com/resource-manager/reference/rest/v1/projects).

Since a FirebaseProject is actually also a GCP Project, a FirebaseProject uses underlying GCP identifiers (most importantly, the projectId) as its own for easy interop with GCP APIs.

The result of this call is an Operation. Poll the Operation to track the provisioning process by calling GetOperation until done is true. When done is true, the Operation has either succeeded or failed. If the Operation succeeded, its response is set to a FirebaseProject; if the Operation failed, its error is set to a google.rpc.Status. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

This method does not modify any billing account information on the underlying GCP Project.

All fields listed in the request body are required.

To call AddFirebase, a member must be an Editor or Owner for the existing GCP Project. Service accounts cannot call AddFirebase.

Arguments

request - No description provided.

project - The resource name of the GCP Project to which Firebase resources will be added, in the format:
projects/projectId After calling AddFirebase, the projectId of the GCP Project is also the projectId of the FirebaseProject.

pub fn search_apps(&self, parent: &str) -> ProjectSearchAppCall<'a, C, A>

Create a builder to help you perform the following task:

A convenience method that lists all available Apps for the specified FirebaseProject.

Typically, interaction with an App should be done using the platform-specific service, but some tool use-cases require a summary of all known Apps (such as for App selector interfaces).

Arguments

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn ios_apps_list(&self, parent: &str) -> ProjectIosAppListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists each IosApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

Arguments

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn web_apps_patch(
&self,
request: WebApp,
name: &str
) -> ProjectWebAppPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the attributes of the WebApp identified by the specified resource name.

Arguments

request - No description provided.

name - The fully qualified resource name of the App, for example:
projects/projectId/webApps/appId

pub fn web_apps_list(&self, parent: &str) -> ProjectWebAppListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists each WebApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

Arguments

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn ios_apps_get(&self, name: &str) -> ProjectIosAppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the IosApp identified by the specified resource name.

Arguments

name - The fully qualified resource name of the App, in the format: projects/projectId/iosApps/appId
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/iosApps/appId

pub fn web_apps_create(
&self,
request: WebApp,
parent: &str
) -> ProjectWebAppCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Requests that a new WebApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

Arguments

request - No description provided.

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn web_apps_get_config(
&self,
name: &str
) -> ProjectWebAppGetConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration artifact associated with the specified WebApp.

Arguments

name - The resource name of the App configuration to download, in the format:
projects/projectId/webApps/appId/config
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/webApps/appId

pub fn android_apps_get_config(
&self,
name: &str
) -> ProjectAndroidAppGetConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration artifact associated with the specified AndroidApp.

Arguments

name - The resource name of the App configuration to download, in the format:
projects/projectId/androidApps/appId/config
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/androidApps/appId

pub fn available_locations_list(
&self,
parent: &str
) -> ProjectAvailableLocationListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of valid Google Cloud Platform (GCP) resource locations for the specified Project (including a FirebaseProject).

The default GCP resource location of a project defines the geographical location where project resources, such as Cloud Firestore, will be provisioned by default.

The returned list are the available GCP resource locations.

This call checks for any location restrictions for the specified Project and, thus, might return a subset of all possible GCP resource locations. To list all GCP resource locations (regardless of any restrictions), call the endpoint without specifying a projectId (that is, /v1beta1/{parent=projects/-}/listAvailableLocations).

To call ListAvailableLocations with a specified project, a member must be at minimum a Viewer of the project. Calls without a specified project do not require any specific project permissions.

Arguments

parent - The Project for which to list GCP resource locations, in the format:
projects/projectId
If no project is specified (that is, projects/-), the returned list does not take into account org-specific or project-specific location restrictions.

pub fn ios_apps_create(
&self,
request: IosApp,
parent: &str
) -> ProjectIosAppCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Requests that a new IosApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

Arguments

request - No description provided.

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn android_apps_sha_delete(
&self,
name: &str
) -> ProjectAndroidAppShaDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a SHA certificate from the specified AndroidApp.

Arguments

name - The fully qualified resource name of the sha-key, in the format:
projects/projectId/androidApps/appId/sha/shaId
You can obtain the full name from the response of ListShaCertificates or the original CreateShaCertificate.

pub fn web_apps_get(&self, name: &str) -> ProjectWebAppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the WebApp identified by the specified resource name.

Arguments

name - The fully qualified resource name of the App, in the format:
projects/projectId/webApps/appId
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/webApps/appId

pub fn android_apps_sha_create(
&self,
request: ShaCertificate,
parent: &str
) -> ProjectAndroidAppShaCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a SHA certificate to the specified AndroidApp.

Arguments

request - No description provided.

parent - The parent App to which a SHA certificate will be added, in the format:
projects/projectId/androidApps/appId
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/androidApps/appId

pub fn patch(
&self,
request: FirebaseProject,
name: &str
) -> ProjectPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the attributes of the FirebaseProject identified by the specified resource name.

All query parameters are required.

Arguments

request - No description provided.

name - The fully qualified resource name of the Project, in the format:
projects/projectId

pub fn android_apps_create(
&self,
request: AndroidApp,
parent: &str
) -> ProjectAndroidAppCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Requests that a new AndroidApp be created.

The result of this call is an Operation which can be used to track the provisioning process. The Operation is automatically deleted after completion, so there is no need to call DeleteOperation.

Arguments

request - No description provided.

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn get_admin_sdk_config(
&self,
name: &str
) -> ProjectGetAdminSdkConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration artifact used by servers to simplify initialization.

Typically, this configuration is used with the Firebase Admin SDK initializeApp command.

Arguments

name - The fully qualified resource name of the Project, in the format:
projects/projectId/adminSdkConfig

pub fn android_apps_list(
&self,
parent: &str
) -> ProjectAndroidAppListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists each AndroidApp associated with the specified parent Project.

The elements are returned in no particular order, but will be a consistent view of the Apps when additional requests are made with a pageToken.

Arguments

parent - The parent Project for which to list Apps, in the format:
projects/projectId

pub fn ios_apps_get_config(
&self,
name: &str
) -> ProjectIosAppGetConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the configuration artifact associated with the specified IosApp.

Arguments

name - The resource name of the App configuration to download, in the format:
projects/projectId/iosApps/appId/config
As an appId is a unique identifier, the Unique Resource from Sub-Collection access pattern may be used here, in the format:
projects/-/iosApps/appId